Update store_ip shell script to check primary interface
This commit updates the store_ip shell script used to get the primary IP address to dynamically determine the primary interface. Without this change the script fails to get the IP address for a CentOS host. This has been tested on CentOS and Debian and provides the primary IP address in both cases.